+++ src/main/java/org/klimenko/Main.java
@@ -1,5 +1,6 @@
 package org.klimenko;
 
+import org.klimenko.controller.MyBot;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
