+++ src/main/java/org/klimenko/Person.java
@@ -3,9 +3,9 @@ package org.klimenko;
 import java.math.BigDecimal;
 
 public class Person {
-    String debtor;
-    String creditor;
-    Double money;
+    public String debtor;
+    public String creditor;
+    public Double money;
 
     public Person(String debtor, String creditor, Double money){
         this.debtor = debtor;
