+++ src/main/java/org/klimenko/SqlProperties.java
@@ -1,7 +1,7 @@
 package org.klimenko;
 
 public class SqlProperties {
-    public final static String url = "jdbc:mysql://192.168.86.51:3306/telegrambot?serverTimezone=America/Los_Angeles&useSSL=false";
-    public final static String user = "telegramuser";
-    public final static String password = "admin1234";
+    public final static String url = "jdbc:mysql://192.168.86.51:3306/telegram?serverTimezone=America/Los_Angeles&useSSL=false&allowPublicKeyRetrieval=true";
+    public final static String user = "user";
+    public final static String password = "password";
 }
