+++ src/main/java/org/klimenko/controller/commands/IPaidYou.java
@@ -1,6 +1,10 @@
-package org.klimenko;
+package org.klimenko.controller.commands;
 
+import org.klimenko.service.Calculus;
+import org.klimenko.Parser;
+import org.klimenko.WrongFormatException;
 import org.telegram.telegrambots.extensions.bots.commandbot.commands.BotCommand;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
@@ -9,32 +13,28 @@ import java.lang.reflect.InvocationTargetException;
 import java.math.BigDecimal;
 import java.sql.SQLException;
 
-public class YouPaidMe extends BotCommand {
-    public YouPaidMe() {
-        super("youpaidme", "You paid your debts");
+public class IPaidYou extends BotCommand {
+    public IPaidYou() {
+        super("ipaidyou", "I paid my debts");
     }
 
     @Override
     public void execute(AbsSender absSender, User user, Chat chat, String[] strings) {
-
-//        String SQL = ;
-
-        String creditor = user.getUserName();
-        String debtor;
+        String creditor;
         BigDecimal money;
+        String debtor = user.getUserName();
         try {
-            debtor = (String) Parser.ParsingMoney(strings).get("name");
+            creditor = (String) Parser.ParsingMoney(strings).get("name");
             money = (BigDecimal) Parser.ParsingMoney(strings).get("amount");
         } catch (WrongFormatException e) {
-            throw new RuntimeException(e);
+            // send a message of failure and return
+            SendMessage sendMessage = new SendMessage();
+            return;
         }
 
         System.out.println(creditor);
         System.out.println(debtor);
         System.out.println(money);
-
-        //SqlConnection.SqlConnect();
-
 
         try {
             Calculus.Transactions(debtor, creditor, money);
