+++ src/main/java/org/klimenko/controller/commands/YouOweMe.java
@@ -1,7 +1,10 @@
-package org.klimenko;
+package org.klimenko.controller.commands;
 
+
+import org.klimenko.service.Calculus;
+import org.klimenko.Parser;
+import org.klimenko.WrongFormatException;
 import org.telegram.telegrambots.extensions.bots.commandbot.commands.BotCommand;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
@@ -10,24 +13,25 @@ import java.lang.reflect.InvocationTargetException;
 import java.math.BigDecimal;
 import java.sql.SQLException;
 
-public class IPaidYou extends BotCommand {
-    public IPaidYou() {
-        super("ipaidyou", "I paid my debts");
+public class YouOweMe extends BotCommand {
+
+    public YouOweMe() {
+        super("youoweme", "You owe me money");
     }
 
     @Override
     public void execute(AbsSender absSender, User user, Chat chat, String[] strings) {
+
+        String debtor = user.getUserName();
         String creditor;
         BigDecimal money;
-        String debtor = user.getUserName();
         try {
             creditor = (String) Parser.ParsingMoney(strings).get("name");
             money = (BigDecimal) Parser.ParsingMoney(strings).get("amount");
         } catch (WrongFormatException e) {
-            // send a message of failure and return
-            SendMessage sendMessage = new SendMessage();
-            return;
+            throw new RuntimeException(e);
         }
+
 
         System.out.println(creditor);
         System.out.println(debtor);
