+++ src/main/java/org/klimenko/controller/commands/YouPaidMe.java
@@ -1,6 +1,8 @@
-package org.klimenko;
+package org.klimenko.controller.commands;
 
-
+import org.klimenko.service.Calculus;
+import org.klimenko.Parser;
+import org.klimenko.WrongFormatException;
 import org.telegram.telegrambots.extensions.bots.commandbot.commands.BotCommand;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
@@ -10,15 +12,15 @@ import java.lang.reflect.InvocationTargetException;
 import java.math.BigDecimal;
 import java.sql.SQLException;
 
-
-public class IOweYou extends BotCommand {
-
-    public IOweYou() {
-        super("ioweyou", "I owe you money");
+public class YouPaidMe extends BotCommand {
+    public YouPaidMe() {
+        super("youpaidme", "You paid your debts");
     }
 
     @Override
     public void execute(AbsSender absSender, User user, Chat chat, String[] strings) {
+
+//        String SQL = ;
 
         String creditor = user.getUserName();
         String debtor;
@@ -29,27 +31,16 @@ public class IOweYou extends BotCommand {
         } catch (WrongFormatException e) {
             throw new RuntimeException(e);
         }
-        System.out.println(debtor);
+
         System.out.println(creditor);
+        System.out.println(debtor);
         System.out.println(money);
 
+        //SqlConnection.SqlConnect();
+
+
         try {
-            if (DAO.CheckTinId(Math.toIntExact(user.getId()))) {
-                if (DAO.CheckTinUsername(debtor)) {
-                    try {
-                        Calculus.Transactions(debtor, creditor, money);
-                    } catch (SQLException | ClassNotFoundException | InvocationTargetException | NoSuchMethodException |
-                             InstantiationException | IllegalAccessException e) {
-                        throw new RuntimeException(e);
-                    }
-                }
-                else{
-                    //вывести что пользователя нет в базе. попросите его зарегестрироваться
-                }
-            }
-            else{
-                DAO.AddToTin(Math.toIntExact(user.getId()), user.getUserName());
-            }
+            Calculus.Transactions(debtor, creditor, money);
         } catch (SQLException | ClassNotFoundException | InvocationTargetException | NoSuchMethodException |
                  InstantiationException | IllegalAccessException e) {
             throw new RuntimeException(e);
